requirements.txt
Django==2.1.2
django-admin==1.3.2
django-excel-response2==2.0.8
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.8.2
gunicorn==19.9.0
mysqlclient==1.3.13
pytz==2018.5
screen==1.0.1
xlwt==1.3.0
